requirements.txt
dominate==2.1.16
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Markdown==0.3
Flask-Moment==0.5.1
Flask-PageDown==0.2.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
ForgeryPy==0.1
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.5
MarkupSafe==0.23
python-slugify==1.2.0
SQLAlchemy==1.0.11
Unidecode==0.4.19
visitor==0.1.2
Werkzeug==0.11.3
WTForms==2.1
